--- app/views/propertyDetails/propertyDetailsTaxAvoidance.scala.html
@@ -1,128 +1,0 @@
-@*
- * Copyright 2023 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import models._
-@import config.ApplicationConfig
-@import play.twirl.api.HtmlFormat
-@import uk.gov.hmrc.govukfrontend.views.html.components._
-@import uk.gov.hmrc.govukfrontend.views.html.components.implicits._
-@import views.ViewUtils.titleBuilder
-@import views.html.helpers._
-@import _root_.utils._
-
-@this(newMain: newMain,
-      govukRadios : GovukRadios,
-      govukInput : GovukInput,
-      govukButton : GovukButton,
-        govukDetails : GovukDetails,
-      govukErrorSummary : GovukErrorSummary,
-      govukBackLink : GovukBackLink,
-      formHelper: FormWithCSRF)
-
-@(id: String, periodKey: Int, propertyDetailsForm: Form[PropertyDetailsTaxAvoidance], mode: Option[String], serviceInfoContent: Html = HtmlFormat.empty, backLink: Option[String])(implicit authContext: StandardAuthRetrievals, messages: Messages, request: Request[AnyContent], appConfig: ApplicationConfig)
-
-@backLinkHtml = {
-    @if(backLink.isDefined) {
-        @govukBackLink(BackLink(
-            href = backLink.get,
-            content = Text("Back")
-        ))
-    }
-}
-
-@paragraphContent = {
-    <p class="govuk-body">@messages("ated.choose-reliefs.avoidance-reveal-line-1")</p>
-    <p class="govuk-body">@messages("ated.choose-reliefs.avoidance-reveal-line-2")</p>
-    <ul class="govuk-list govuk-list--bullet">
-        <li>@messages("ated.choose-reliefs.avoidance-info-line-1")</li>
-        <li>@messages("ated.choose-reliefs.avoidance-info-line-2")</li>
-    </ul>
-    <a class="govuk-link govuk-body" href="https://www.gov.uk/guidance/disclosure-of-tax-avoidance-schemes-overview" rel="noreferrer noopener" target="_blank">@Html(messages("ated.choose-reliefs.avoidance-more-info"))</a>
-}
-
-@avoidanceSchemeHTML = {
-    @govukInput(Input(
-        label = Label(
-            content = Text(messages("ated.property-details-period.taxAvoidanceScheme"))
-        ),
-        hint = Some(Hint(
-            content = Text(messages("ated.choose-reliefs.avoidance-hint"))
-        ))
-    ).withFormField(propertyDetailsForm("taxAvoidanceScheme")))
-
-    @govukInput(Input(
-        label = Label(
-            content = Text(messages("ated.property-details-period.taxAvoidancePromoterReference"))
-        ),
-        hint = Some(Hint(
-            content = Text(messages("ated.choose-reliefs.avoidance-hint"))
-        ))
-    ).withFormField(propertyDetailsForm("taxAvoidancePromoterReference")))
-}
-
-@newMain(title = titleBuilder(messages("ated.property-details-period.isTaxAvoidance.title"), Some(propertyDetailsForm)),
-          serviceInfoContent = serviceInfoContent) {
-
-    @backLinkHtml
-
-    @if(propertyDetailsForm.errors.nonEmpty) {
-        @govukErrorSummary(ErrorSummary().withFormErrorsAsText(propertyDetailsForm))
-    }
-
-   <header>
-        <h2 class="govuk-caption-xl hmrc-caption-xl">
-            <span class="govuk-visually-hidden">@messages("ated.screen-reader.section")</span>
-            @messages(AtedUtils.getPropertyDetailsPreHeader(mode))
-        </h2>
-        <h1 class="govuk-heading-xl">
-          @messages("ated.property-details-period.isTaxAvoidance.header")
-        </h1>
-   </header>
-
-    @govukDetails(Details(
-        summary = Text(messages("ated.choose-reliefs.avoidance-question")),
-        content = HtmlContent(paragraphContent)
-    ))
-
-@formHelper(action=controllers.propertyDetails.routes.PropertyDetailsTaxAvoidanceController.save(id, periodKey, mode)) {
-  
-  @govukRadios(Radios(
-    fieldset = Some(Fieldset(
-        legend = Some(Legend(
-            content = Text(messages("ated.property-details-period.isTaxAvoidance.header")),
-            classes = "govuk-visually-hidden"
-         ))
-    )),
-    items = Seq(
-      RadioItem(
-        content = Text(messages("ated.property-details-value.yes")), 
-        value = Some("true"),
-        conditionalHtml = Some(avoidanceSchemeHTML)
-      ), 
-      RadioItem(
-        content = Text(messages("ated.property-details-value.no")),
-        value = Some("false")))
-      ).withFormField(propertyDetailsForm("isTaxAvoidance"))
-    )
-
-    @govukButton(Button(
-        content = Text(messages("ated.save-and-continue")),
-        attributes = Map("id" -> "submit")
-    ))
-
-  }
-
-}
