+++ app/views/propertyDetails/propertyDetailsTaxAvoidanceReferences.scala.html
@@ -63,7 +63,7 @@
    </header>
 
    
-@formHelper(action=controllers.test.routes.PropertyDetailsTaxAvoidanceReferencesController.save(id, periodKey, mode)) {
+@formHelper(action=controllers.propertyDetails.routes.PropertyDetailsTaxAvoidanceReferencesController.save(id, periodKey, mode)) {
   
   @govukInput(Input(
         label = Label(
