+++ conf/messages
@@ -725,7 +725,7 @@ ated.property-details-period.isTaxAvoidanceScheme.title = Is an avoidance scheme
 ated.property-details-period.isTaxAvoidanceScheme.header = Is an avoidance scheme being used?
 
 ated.property-details-period.isTaxAvoidance = Is an avoidance scheme being used?
-ated.property-details-period.isTaxAvoidance.error-field-name = Select yes if a tax avoidance scheme is being used
+ated.property-details-period.isTaxAvoidance.error-field-name = You must select either Yes or No
 ated.property-details-period.isTaxAvoidanceScheme.error-field-name = You must select either Yes or No
 ated.property-details-period.taxAvoidanceScheme = Avoidance scheme reference number
 ated.property-details-period.taxAvoidanceScheme.error.empty = Enter a valid avoidance scheme reference number
