+++ conf/testOnlyDoNotUseInAppConf.routes
@@ -11,11 +11,3 @@
 
 # Add all the application routes to the prod.routes file
 ->         /                          prod.Routes
-
-GET      /test-only/liability/create/tax-avoidance-reference/edit-summary/:propertyKey             controllers.test.PropertyDetailsTaxAvoidanceReferencesController.editFromSummary(propertyKey: String)
-GET      /test-only/liability/create/tax-avoidance-reference/view/:propertyKey                     controllers.test.PropertyDetailsTaxAvoidanceReferencesController.view(propertyKey: String)
-POST     /test-only/liability/create/tax-avoidance-reference/save/:propertyKey/period/:periodKey   controllers.test.PropertyDetailsTaxAvoidanceReferencesController.save(propertyKey: String, periodKey: Int, mode: Option[String] ?=None)
-
-GET      /test-only/liability/create/tax-avoidance/edit-summary/:propertyKey             controllers.test.PropertyDetailsTaxAvoidanceSchemeController.editFromSummary(propertyKey: String)
-GET      /test-only/liability/create/tax-avoidance/view/:propertyKey                     controllers.test.PropertyDetailsTaxAvoidanceSchemeController.view(propertyKey: String)
-POST     /test-only/liability/create/tax-avoidance/save/:propertyKey/period/:periodKey   controllers.test.PropertyDetailsTaxAvoidanceSchemeController.save(propertyKey: String, periodKey: Int, mode: Option[String] ?=None)
