--- test/controllers/propertyDetails/PropertyDetailsTaxAvoidanceControllerSpec.scala
@@ -1,281 +1,0 @@
-/*
- * Copyright 2023 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package controllers.propertyDetails
-
-import java.util.UUID
-
-import builders.{PropertyDetailsBuilder, SessionBuilder, TitleBuilder}
-import config.ApplicationConfig
-import connectors.{BackLinkCacheConnector, DataCacheConnector}
-import controllers.auth.AuthAction
-import models._
-import java.time.LocalDate
-import org.jsoup.Jsoup
-import org.mockito.ArgumentMatchers
-import org.mockito.Mockito._
-import org.scalatest.BeforeAndAfterEach
-import org.scalatestplus.mockito.MockitoSugar
-import org.scalatestplus.play.PlaySpec
-import org.scalatestplus.play.guice.GuiceOneServerPerSuite
-import play.api.i18n.{Lang, MessagesApi, MessagesImpl}
-import play.api.libs.json.{JsValue, Json}
-import play.api.mvc.{MessagesControllerComponents, Result}
-import play.api.test.FakeRequest
-import play.api.test.Helpers._
-import services.{PropertyDetailsCacheSuccessResponse, PropertyDetailsService, ServiceInfoService, SubscriptionDataService}
-import testhelpers.MockAuthUtil
-import uk.gov.hmrc.auth.core.AffinityGroup
-import uk.gov.hmrc.http.HeaderCarrier
-import utils.{AtedConstants, PeriodUtils}
-import views.html.BtaNavigationLinks
-
-import scala.concurrent.Future
-
-
-class PropertyDetailsTaxAvoidanceControllerSpec extends PlaySpec with GuiceOneServerPerSuite with MockitoSugar with BeforeAndAfterEach with MockAuthUtil {
-  implicit val mockAppConfig: ApplicationConfig = app.injector.instanceOf[ApplicationConfig]
-  implicit lazy val hc: HeaderCarrier = HeaderCarrier()
-  lazy implicit val messages: MessagesImpl = MessagesImpl(Lang("en-GB"), messagesApi)
-
-  val mockMcc: MessagesControllerComponents = app.injector.instanceOf[MessagesControllerComponents]
-  val mockPropertyDetailsService: PropertyDetailsService = mock[PropertyDetailsService]
-  val mockDataCacheConnector: DataCacheConnector = mock[DataCacheConnector]
-  val mockBackLinkCacheConnector: BackLinkCacheConnector = mock[BackLinkCacheConnector]
-  val mockSubscriptionDataService: SubscriptionDataService = mock[SubscriptionDataService]
-  val mockPropertyDetailsSupportingInfoController: PropertyDetailsSupportingInfoController = mock[PropertyDetailsSupportingInfoController]
-  val messagesApi: MessagesApi = app.injector.instanceOf[MessagesApi]
-  val btaNavigationLinksView: BtaNavigationLinks = app.injector.instanceOf[BtaNavigationLinks]
-  val mockServiceInfoService: ServiceInfoService = mock[ServiceInfoService]
-  val injectedViewInstance = app.injector.instanceOf[views.html.propertyDetails.propertyDetailsTaxAvoidance]
-
-  val periodKey: Int = PeriodUtils.calculatePeakStartYear()
-
-  class Setup {
-
-    val mockAuthAction: AuthAction = new AuthAction(
-      mockAppConfig,
-      mockDelegationService,
-      mockAuthConnector
-    )
-
-    val testPropertyDetailsTaxAvoidanceController: PropertyDetailsTaxAvoidanceController = new PropertyDetailsTaxAvoidanceController(
-      mockMcc,
-      mockAuthAction,
-      mockPropertyDetailsSupportingInfoController,
-      mockServiceInfoService,
-      mockPropertyDetailsService,
-      mockDataCacheConnector,
-      mockBackLinkCacheConnector,
-      injectedViewInstance
-    )
-
-    def getWithUnAuthorisedUser(test: Future[Result] => Any): Unit = {
-      val userId = s"user-${UUID.randomUUID}"
-      val authMock = authResultDefault(AffinityGroup.Organisation, invalidEnrolmentSet)
-      setInvalidAuthMocks(authMock)
-      val result = testPropertyDetailsTaxAvoidanceController.view("1").apply(SessionBuilder.buildRequestWithSession(userId))
-      test(result)
-    }
-
-    def getDataWithAuthorisedUser(propertyDetails: PropertyDetails)(test: Future[Result] => Any): Unit = {
-      val userId = s"user-${UUID.randomUUID}"
-      val authMock = authResultDefault(AffinityGroup.Organisation, defaultEnrolmentSet)
-      setAuthMocks(authMock)
-      when(mockServiceInfoService.getPartial(ArgumentMatchers.any(), ArgumentMatchers.any(), ArgumentMatchers.any())).thenReturn(Future.successful(btaNavigationLinksView()(messages,mockAppConfig)))
-      when(mockDataCacheConnector.fetchAtedRefData[String](ArgumentMatchers.eq(AtedConstants.DelegatedClientAtedRefNumber))
-        (ArgumentMatchers.any(), ArgumentMatchers.any())).thenReturn(Future.successful(Some("XN1200000100001")))
-      when(mockDataCacheConnector.fetchAndGetFormData[Boolean](ArgumentMatchers.any())
-        (ArgumentMatchers.any(), ArgumentMatchers.any())).thenReturn(Future.successful(None))
-      when(mockBackLinkCacheConnector.fetchAndGetBackLink(ArgumentMatchers.any())(ArgumentMatchers.any())).thenReturn(Future.successful(None))
-      when(mockPropertyDetailsService.retrieveDraftPropertyDetails(ArgumentMatchers.any())(ArgumentMatchers.any(), ArgumentMatchers.any()))
-        .thenReturn(Future.successful(PropertyDetailsCacheSuccessResponse(propertyDetails)))
-      val result = testPropertyDetailsTaxAvoidanceController.view(propertyDetails.id).apply(SessionBuilder.buildRequestWithSession(userId))
-      test(result)
-    }
-
-    def editFromSummary(propertyDetails: PropertyDetails)(test: Future[Result] => Any): Unit = {
-      val userId = s"user-${UUID.randomUUID}"
-      val authMock = authResultDefault(AffinityGroup.Organisation, defaultEnrolmentSet)
-      setAuthMocks(authMock)
-      when(mockDataCacheConnector.fetchAtedRefData[String](ArgumentMatchers.eq(AtedConstants.DelegatedClientAtedRefNumber))
-        (ArgumentMatchers.any(), ArgumentMatchers.any())).thenReturn(Future.successful(Some("XN1200000100001")))
-      when(mockPropertyDetailsService.retrieveDraftPropertyDetails(ArgumentMatchers.any())(ArgumentMatchers.any(), ArgumentMatchers.any()))
-        .thenReturn(Future.successful(PropertyDetailsCacheSuccessResponse(propertyDetails)))
-      val result = testPropertyDetailsTaxAvoidanceController.editFromSummary(propertyDetails.id).apply(SessionBuilder.buildRequestWithSession(userId))
-      test(result)
-    }
-
-    def saveWithUnAuthorisedUser(test: Future[Result] => Any): Unit = {
-      val periodKey: Int = 2015
-      val userId = s"user-${UUID.randomUUID}"
-      val authMock = authResultDefault(AffinityGroup.Organisation, invalidEnrolmentSet)
-      setInvalidAuthMocks(authMock)
-      val result = testPropertyDetailsTaxAvoidanceController.save("1", periodKey, None).apply(SessionBuilder.buildRequestWithSession(userId))
-      test(result)
-    }
-
-    def submitWithAuthorisedUser(inputJson: JsValue)(test: Future[Result] => Any): Unit = {
-      val periodKey: Int = 2015
-      val userId = s"user-${UUID.randomUUID}"
-      when(mockDataCacheConnector.fetchAtedRefData[String](ArgumentMatchers.eq(AtedConstants.DelegatedClientAtedRefNumber))
-        (ArgumentMatchers.any(), ArgumentMatchers.any())).thenReturn(Future.successful(Some("XN1200000100001")))
-      when(mockPropertyDetailsService.saveDraftPropertyDetailsTaxAvoidance(ArgumentMatchers.eq("1"), ArgumentMatchers.any())(ArgumentMatchers.any(), ArgumentMatchers.any())).
-        thenReturn(Future.successful(OK))
-      val authMock = authResultDefault(AffinityGroup.Organisation, defaultEnrolmentSet)
-      setAuthMocks(authMock)
-      val result = testPropertyDetailsTaxAvoidanceController.save("1", periodKey, None)
-        .apply(SessionBuilder.updateRequestWithSession(FakeRequest().withJsonBody(inputJson), userId))
-      test(result)
-    }
-  }
-
-  override def beforeEach(): Unit = {
-  }
-
-  "PropertyDetailsTaxAvoidanceController" must {
-    "propertyDetails" must {
-
-      "unauthorised users" must {
-
-        "respond with a redirect" in new Setup {
-          getWithUnAuthorisedUser { result =>
-            status(result) must be(SEE_OTHER)
-          }
-        }
-
-        "be redirected to the unauthorised page" in new Setup {
-          getWithUnAuthorisedUser { result =>
-            redirectLocation(result).get must include("/ated/unauthorised")
-          }
-        }
-      }
-
-      "Authorised users" must {
-
-        "show the chargeable property details value view with no data" in new Setup {
-          val propertyDetails: PropertyDetails = PropertyDetailsBuilder.getPropertyDetails("1", Some("postCode")).copy(period = None)
-
-          getDataWithAuthorisedUser(propertyDetails) {
-            result =>
-              status(result) must be(OK)
-              val document = Jsoup.parse(contentAsString(result))
-              document.title() must be(TitleBuilder.buildTitle("Is a tax avoidance scheme being used?"))
-              document.getElementsByAttributeValue("for","isTaxAvoidance").text() mustBe messages("ated.property-details-value.yes")
-              document.getElementsByAttributeValue("for","isTaxAvoidance-2").text() mustBe messages("ated.property-details-value.no")
-              assert(document.getElementById("service-info-list").text() === "Home Manage account Messages Help and contact")
-
-              document.getElementById("taxAvoidanceScheme").attr("value") must be("")
-
-              document.getElementsByClass("govuk-details__text")
-                .text() must include("HMRC never approves tax avoidance schemes. You must tell us if you are using a tax avoidance scheme that falls within the Disclosure of Tax Avoidance Schemes (DOTAS).")
-              document.getElementsByClass("govuk-details__text")
-                .text() must include("You will have received your scheme reference number (SRN) and promoter reference number (PRN)")
-
-              document.getElementsByClass("govuk-button").text() must be("Save and continue")
-          }
-        }
-
-        "show the chargeable property details value view with existing data" in new Setup {
-
-          val propertyDetailsPeriod: Option[PropertyDetailsPeriod] = PropertyDetailsBuilder.getPropertyDetailsPeriodDatesLiable(LocalDate.parse("1970-12-01"), LocalDate.parse("1999-03-02")).
-            map(_.copy(taxAvoidanceScheme = Some("taxAvoid")))
-
-          val propertyDetails: PropertyDetails = PropertyDetailsBuilder.getPropertyDetails("1", Some("postCode")).copy(period = propertyDetailsPeriod)
-          getDataWithAuthorisedUser(propertyDetails) {
-            result =>
-              status(result) must be(OK)
-              val document = Jsoup.parse(contentAsString(result))
-              document.getElementById("isTaxAvoidance").attr("value") must be("true")
-              document.getElementById("isTaxAvoidance-2").attr("value") must be("false")
-              document.getElementById("taxAvoidanceScheme").attr("value") must be("taxAvoid")
-
-              document.getElementsByClass("govuk-button").text() must be("Save and continue")
-          }
-        }
-      }
-    }
-
-    "editFromSummary" must {
-
-      "Authorised users" must {
-
-        "show the chargeable property details value view with no data" in new Setup {
-          val propertyDetails: PropertyDetails = PropertyDetailsBuilder.getPropertyDetails("1", Some("postCode")).copy(period = None)
-
-          editFromSummary(propertyDetails) {
-            result =>
-              status(result) must be(OK)
-              val document = Jsoup.parse(contentAsString(result))
-              document.title() must be(TitleBuilder.buildTitle("Is a tax avoidance scheme being used?"))
-
-              document.getElementsByClass("govuk-back-link").text must be("Back")
-              document.getElementsByClass("govuk-back-link").attr("href") must include("/ated/liability/create/summary")
-          }
-        }
-
-        "show the chargeable property details value view with no data with a propertyDetails period" in new Setup {
-          val propertyDetails: PropertyDetails = PropertyDetailsBuilder.getPropertyDetails("1", Some("postCode"))
-
-          editFromSummary(propertyDetails) {
-            result =>
-              status(result) must be(OK)
-              val document = Jsoup.parse(contentAsString(result))
-              document.title() must be(TitleBuilder.buildTitle("Is a tax avoidance scheme being used?"))
-
-              document.getElementsByClass("govuk-back-link").text must be("Back")
-              document.getElementsByClass("govuk-back-link").attr("href") must include("/ated/liability/create/summary")
-          }
-        }
-      }
-    }
-    "save" must {
-      "unauthorised users" must {
-
-        "be redirected to the login page" in new Setup {
-          saveWithUnAuthorisedUser { result =>
-            status(result) must be(SEE_OTHER)
-            redirectLocation(result).get must include("/ated/unauthorised")
-          }
-        }
-      }
-
-      "Authorised users" must {
-        "for invalid data, return BAD_REQUEST" in new Setup {
-          val taxAvoidance: PropertyDetailsTaxAvoidance = PropertyDetailsTaxAvoidance(Some(true))
-
-          when(mockBackLinkCacheConnector.fetchAndGetBackLink(ArgumentMatchers.any())(ArgumentMatchers.any())).thenReturn(Future.successful(None))
-          submitWithAuthorisedUser(Json.toJson(taxAvoidance)) {
-            result =>
-              status(result) must be(BAD_REQUEST)
-          }
-        }
-        "for valid data, return OK" in new Setup {
-          val taxAvoidance: PropertyDetailsTaxAvoidance = PropertyDetailsTaxAvoidance(Some(false))
-
-          when(mockBackLinkCacheConnector.saveBackLink(ArgumentMatchers.any(), ArgumentMatchers.any())(ArgumentMatchers.any()))
-            .thenReturn(Future.successful(None))
-
-          submitWithAuthorisedUser(Json.toJson(taxAvoidance)) {
-            result =>
-              status(result) must be(SEE_OTHER)
-              redirectLocation(result).get must include("/liability/create/supporting-info/view")
-          }
-        }
-      }
-    }
-  }
-}
