+++ test/controllers/propertyDetails/PropertyDetailsTaxAvoidanceSchemeControllerSpec.scala
@@ -20,7 +20,6 @@ import builders.SessionBuilder
 import config.ApplicationConfig
 import connectors.{BackLinkCacheConnector, DataCacheConnector}
 import controllers.auth.AuthAction
-import controllers.test.PropertyDetailsTaxAvoidanceSchemeController
 import models._
 import org.mockito.ArgumentMatchers
 import org.mockito.Mockito._
@@ -40,7 +39,6 @@ import uk.gov.hmrc.http.HeaderCarrier
 import utils.{AtedConstants, PeriodUtils}
 import views.html.BtaNavigationLinks
 
-
 import java.util.UUID
 import scala.concurrent.Future
 
@@ -55,6 +53,7 @@ class PropertyDetailsTaxAvoidanceSchemeControllerSpec extends PlaySpec with Guic
   val mockDataCacheConnector: DataCacheConnector = mock[DataCacheConnector]
   val mockBackLinkCacheConnector: BackLinkCacheConnector = mock[BackLinkCacheConnector]
   val mockSubscriptionDataService: SubscriptionDataService = mock[SubscriptionDataService]
+  val mockPropertyDetailsTaxAvoidanceReferencesController: PropertyDetailsTaxAvoidanceReferencesController = mock[PropertyDetailsTaxAvoidanceReferencesController]
   val mockPropertyDetailsSupportingInfoController: PropertyDetailsSupportingInfoController = mock[PropertyDetailsSupportingInfoController]
   val messagesApi: MessagesApi = app.injector.instanceOf[MessagesApi]
   val btaNavigationLinksView: BtaNavigationLinks = app.injector.instanceOf[BtaNavigationLinks]
@@ -74,6 +73,7 @@ class PropertyDetailsTaxAvoidanceSchemeControllerSpec extends PlaySpec with Guic
     val testPropertyDetailsTaxAvoidanceSchemeController: PropertyDetailsTaxAvoidanceSchemeController = new PropertyDetailsTaxAvoidanceSchemeController(
       mockMcc,
       mockAuthAction,
+      mockPropertyDetailsTaxAvoidanceReferencesController,
       mockPropertyDetailsSupportingInfoController,
       mockServiceInfoService,
       mockPropertyDetailsService,
